+++ src/directory/chaining_data_store.h
@@ -162,7 +162,7 @@ public:
         const auto alloc = allocate (data.size());
 
         unsigned long offset = 0;
-        int index = 0;
+        size_t index = 0;
         size_t total_acquired_size = 0;
         for (const auto& partial_alloc: alloc) {
             if (lseek(partial_alloc.fd, partial_alloc.seek, SEEK_SET) != partial_alloc.seek) [[unlikely]] {
@@ -180,7 +180,7 @@ public:
             }
             const auto header_buffer = h.serialize();
             const auto ws = ::write (partial_alloc.fd, header_buffer.data(), header_size);
-            if (ws != header_size) [[unlikely]] {
+            if (ws != static_cast <ssize_t> (header_size)) [[unlikely]] {
                 throw std::runtime_error ("Could not write the header in the chaining data store");
             }
             for (size_t written = 0;
